+++ drivers/src/hpm_i2c_drv.c
@@ -139,7 +139,7 @@ hpm_stat_t i2c_init_master(I2C_Type *ptr, uint32_t src_clk_in_hz, i2c_config_t *
 }
 
 hpm_stat_t i2c_master_address_read(I2C_Type *ptr, const uint16_t device_address,
-                                   uint8_t *addr, uint8_t addr_size_in_byte,
+                                   uint8_t *addr, uint32_t addr_size_in_byte,
                                    uint8_t *buf, const uint32_t size_in_byte)
 {
     hpm_stat_t stat = status_success;
@@ -222,7 +222,7 @@ hpm_stat_t i2c_master_address_read(I2C_Type *ptr, const uint16_t device_address,
 }
 
 hpm_stat_t i2c_master_address_write(I2C_Type *ptr, const uint16_t device_address,
-                                    uint8_t *addr, uint8_t addr_size_in_byte,
+                                    uint8_t *addr, uint32_t addr_size_in_byte,
                                     uint8_t *buf, const uint32_t size_in_byte)
 {
     hpm_stat_t stat = status_success;
